move version and variant logic to pkg
